afxdp/cos/flow_hash: relocate inline tests to flow_hash_tests.rs
Same playbook (#1080-#1112). flow_hash.rs: 425 → 174 LOC.
cargo build/test 92/878.
